Day 8

+++ 2023/cpp/src/Day.h
@@ -10,9 +10,9 @@ public:
 
     virtual ~Day() = default;
 
-    virtual int part1() = 0;
+    virtual long part1() = 0;
 
-    virtual int part2() = 0;
+    virtual long part2() = 0;
 
     [[nodiscard]] int getDay() const;
