+++ 2023/cpp/src/day1/Day1.h
@@ -12,9 +12,9 @@ public:
 
     ~Day1() override;
 
-    int part1() override;
+    long part1() override;
 
-    int part2() override;
+    long part2() override;
 };
 
 #endif /* Day1_h */
