+++ 2023/cpp/src/day3/Day3.h
@@ -9,9 +9,9 @@ public:
 
     ~Day3() override;
 
-    int part1() override;
+    long part1() override;
 
-    int part2() override;
+    long part2() override;
 };
 
 #endif /* Day3_h */
