+++ 2023/cpp/src/day4/Day4.h
@@ -9,9 +9,9 @@ public:
 
     ~Day4() override;
 
-    int part1() override;
+    long part1() override;
 
-    int part2() override;
+    long part2() override;
 };
 
 #endif /* Day4_h */
