+++ 2023/cpp/src/day6/Day6.h
@@ -9,9 +9,9 @@ public:
 
     ~Day6() override;
 
-    int part1() override;
+    long part1() override;
 
-    int part2() override;
+    long part2() override;
 
 private:
     static long getDistanceForHoldTime(int time, int holdTime);
